+++ grth/src/Akeneo/Platform/Bundle/MonitoringBundle/ServiceStatusChecker/PubSub/PubSubStatusChecker.php
@@ -7,6 +7,7 @@ namespace Akeneo\Platform\Bundle\MonitoringBundle\ServiceStatusChecker\PubSub;
 use Akeneo\Platform\Bundle\MonitoringBundle\ServiceStatusChecker\ServiceStatus;
 use Akeneo\Tool\Bundle\MessengerBundle\Transport\GooglePubSub\PubSubClientFactory;
 use Google\Cloud\Core\Exception\ServiceException;
+use Psr\Log\LoggerInterface;
 
 /**
  * @author    Thomas Galvaing <thomas.galvaing@akeneo.com>
@@ -19,27 +20,28 @@ class PubSubStatusChecker implements PubSubStatusCheckerInterface
     private string $projectId;
     private string $topicName;
     private string $subscriptionName;
+    private LoggerInterface $logger;
 
     public function __construct(
         PubSubClientFactory $pubSubClientFactory,
         string $projectId,
         string $topicName,
-        string $subscriptionName
+        string $subscriptionName,
+        LoggerInterface $logger
     ) {
         $this->pubSubClientFactory = $pubSubClientFactory;
         $this->projectId = $projectId;
         $this->topicName = $topicName;
         $this->subscriptionName = $subscriptionName;
+        $this->logger = $logger;
     }
 
     public function status(): ServiceStatus
     {
-        $pubSubClient = $this->pubSubClientFactory->createPubSubClient(['projectId' => $this->projectId]);
-
-        $topic = $pubSubClient->topic($this->topicName);
-        $subscription = $topic->subscription($this->subscriptionName);
-
         try {
+            $pubSubClient = $this->pubSubClientFactory->createPubSubClient(['projectId' => $this->projectId]);
+            $topic = $pubSubClient->topic($this->topicName);
+            $subscription = $topic->subscription($this->subscriptionName);
             $messages = $subscription->pull([
                 'maxMessages' => 1,
                 'returnImmediately' => true,
@@ -48,7 +50,8 @@ class PubSubStatusChecker implements PubSubStatusCheckerInterface
             if (count($messages) > 0) {
                 $subscription->modifyAckDeadline($messages[0], 0);
             }
-        } catch (ServiceException $exception) {
+        } catch (\Throwable $exception) {
+            $this->logger->error("PubSub ServiceCheck error", ['exception' => $exception]);
             return ServiceStatus::notOk(sprintf('Unable to access Pub/Sub: %s', $exception->getMessage()));
         }