+++ CHANGELOG.md
@@ -104,6 +104,7 @@
 - PIM-10087: Fix storage errors HTTP code to return 500 instead of 422
 - PIM-10048: fix memory leak in search product models by family variant query
 - PIM-10115: Blacklist domains for asset media links
+- PIM-10110: Enrich Service check logging upon exception and slow service check.
 
 ## Improvements